requirements.txt
pandas>=1.1.1
boto3>=1.34.44
Pillow>=8.1.0